services/api/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
scikit-image==0.22.0
numpy==1.26.2
PyGithub==2.1.1
opencv-python==4.8.1.78
gunicorn==21.2.0
google-generativeai>=0.8.0
